# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Real-time Chat API",
        "description": "Real-time chat application API",
        "contact": {
            "email": "contact@test.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local de Desenvolvimento"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/chats/{chatId}/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "List messages in a chat",
                "operationId": "9674200f7930a7c481102359f5aab97f",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID do chat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de mensagens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 50
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Busca por conteúdo na mensagem",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "como vai"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de mensagens paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Message"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat não encontrado ou usuário não é participante"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send a new message to a chat",
                "operationId": "4ddee1b7fed4f0fb8b9aa89baec494b2",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID do chat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Esta é uma nova mensagem."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "$ref": "#/components/schemas/Message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat não encontrado ou usuário não é participante"
                    },
                    "422": {
                        "description": "Validation error (e.g., empty content)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chats/{chatId}/messages/{messageId}": {
            "put": {
                "tags": [
                    "Messages"
                ],
                "summary": "Update a message",
                "operationId": "3b40fbf32546d6032556644f21ecce88",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID do chat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "messageId",
                        "in": "path",
                        "description": "ID da mensagem",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Esta é a mensagem atualizada."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "$ref": "#/components/schemas/Message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat ou mensagem não encontrado, ou usuário não é o autor"
                    },
                    "422": {
                        "description": "Validation error (e.g., empty content)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Delete a message",
                "operationId": "33d8e2f8df5e3e2c2c8d5c575c1df04b",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID do chat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "messageId",
                        "in": "path",
                        "description": "ID da mensagem",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Message deleted successfully"
                    },
                    "404": {
                        "description": "Chat ou mensagem não encontrado, ou usuário não é o autor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chats/{chatId}/messages/{messageId}/read": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Mark a message as read",
                "operationId": "d697aed60e27dcdf246942611714f160",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID do chat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "messageId",
                        "in": "path",
                        "description": "ID da mensagem",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message marked as read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Message marked as read"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat ou mensagem não encontrado, ou usuário é o autor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messages/search": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Search messages across all chats",
                "operationId": "2d3311eede42ce8e10a717d72041efc9",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Termo de busca",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "palavra-chave"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de resultados por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results paginated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Message"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error (e.g., missing search term)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Message": {
                "title": "Message",
                "required": [
                    "id",
                    "content",
                    "user_id",
                    "chat_id",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "content": {
                        "type": "string",
                        "example": "This is a message content."
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "chat_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:34:56Z"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            },
                            "avatar": {
                                "type": "string",
                                "format": "url",
                                "example": "http://example.com/avatar.jpg"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}